requirement.txt
google-generativeai
pandas
streamlit
python-dotenv
beautifulsoup4
lxml